remove webpack, header and search almost complete
