--- package.json
@@ -1,29 +1,0 @@
-{
-  "name": "lab-html-exercise",
-  "version": "1.0.0",
-  "private": true,
-  "description": "![logo_ironhack_blue 7](https://user-images.githubusercontent.com/23629340/40541063-a07a0a8a-601a-11e8-91b5-2f13e4e6b441.png)",
-  "main": "main.js",
-  "scripts": {
-    "start": "webpack --config webpack.config.js",
-    "dev": "webpack-dev-server"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/JoFont/lab-html-exercise.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/JoFont/lab-html-exercise/issues"
-  },
-  "homepage": "https://github.com/JoFont/lab-html-exercise#readme",
-  "dependencies": {
-    "all-the-package-names": "^1.3842.0"
-  },
-  "devDependencies": {
-    "webpack": "^4.40.2",
-    "webpack-cli": "^3.3.8",
-    "webpack-dev-server": "^3.8.0"
-  }
-}
