--- static/styles/colors.css
@@ -1,68 +1,0 @@
-:root {
-    
-    /*! Color Variables */
-
-    --_white: 255, 255, 255;
-    --_white-shade: 247, 247, 247;
-
-    --_light: 248, 244, 244;
-
-    --_black: 0, 0, 0;
-    --_black-shade: 51, 51, 51;
-
-
-    --_red: 251, 59, 73;
-    --_red-shade: 204, 4, 19;
-
-    --_red-hot: 204, 4, 19;
-
-    
-    /*! Global Colors */
-
-    --white: rgba(var(--_white), 1);
-    --light: rgba(var(--_light), 1);
-    --black: rgba(var(--_black), 1);
-    --red-hot: rgba(var(--_red-hot), 1);
-    --red: rgba(var(--_red), 1);
-
-}
-
-/** Backgrounds */
-
-    .bg-white {
-        background-color: var(--white);
-    }
-
-    .bg-light {
-        background-color: var(--light);
-    }
-
-
-/*! Misc */
-
-.border-b-sm {
-    border-bottom: solid 1px rgba(var(--_black), 0.1);
-}
-
-.shadow-sm {
-    box-shadow: 0 4px 13px -3px rgba(var(--_black), 0.1);
-}
-
-.pointer {
-    cursor: pointer;
-}
-
-.gradient-border {
-    border-top-width: 3px 0 0;
-    border-top-style: solid;
-
-    -o-border-image: linear-gradient(139deg, #fb8817, #ff4b01, #c12127, #e02aff) 3;
-    border-image: linear-gradient(139deg, #fb8817, #ff4b01, #c12127, #e02aff) 3;
-
-
-    border-image-outset: 0;
-    border-image-repeat: stretch;
-    border-image-slice: 3;
-    border-image-source: linear-gradient(139deg, rgb(251, 136, 23), rgb(255, 75, 1), rgb(193, 33, 39), rgb(224, 42, 255));
-    border-image-width: 1;
-}
