--- static/styles/global.css
@@ -1,49 +1,0 @@
-@import url('https://fonts.googleapis.com/css?family=Poppins:100,100i,200,200i,300,300i,400,400i,500,500i,600,600i,700,700i,800,800i,900,900i&display=swap');
-@import url('https://fonts.googleapis.com/css?family=Fira+Mono:400,500,700&display=swap');
-@import "reset.css";
-@import "colors.css";
-@import "sizes.css";
-@import "components.css";
-
-body {
-    font-family: "Poppins", sans-serif;
-}
-
-
-header {
-    width: 100%;
-
-    font-size: 14px;
-}
-
-
-.header-wraper {
-    display: flex;
-    justify-content: space-around;
-    width: 100%;
-}
-
-
-.header-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    max-width: 93rem;
-}
-
-.npm-expansions {
-    display: inline-flex;
-}
-
-nav ul {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    font-weight: 500;
-}
-
-.brand {
-    max-width: 4.1rem;
-}
-
