+++ styles/sizes.css
@@ -1,0 +1,344 @@
+:root {
+    --_0: 0;
+    --_1: 0.2rem;
+    --_2: 0.4rem;
+    --_3: 0.6rem;
+    --_4: 0.8rem;
+    --_5: 1rem;
+    --_6: 1.2rem;
+    --_7: 2rem;
+    --_8: 2.8rem;
+    --_9: 4rem;
+    --_10: 5rem;
+
+}
+
+/**! Margins */
+
+    /** Margin */
+
+    .m-0 {
+        margin: var(--_0);
+    }
+    .m-1 {
+        margin: var(--_1);
+    }
+    .m-2 {
+        margin: var(--_2);
+    }
+    .m-3 {
+        margin: var(--_3);
+    }
+    .m-4 {
+        margin: var(--_4);
+    }
+    .m-5 {
+        margin: var(--_5);
+    }
+    .m-6 {
+        margin: var(--_6);
+    }
+    .m-7 {
+        margin: var(--_7);
+    }
+    .m-8 {
+        margin: var(--_8);
+    }
+    .m-9 {
+        margin: var(--_9);
+    }
+    .m-10 {
+        margin: var(--_10);
+    }
+
+
+    /** Margin Right */
+
+    .mr-0 {
+        margin-right: var(--_0);
+    }
+    .mr-1 {
+        margin-right: var(--_1);
+    }
+    .mr-2 {
+        margin-right: var(--_2);
+    }
+    .mr-3 {
+        margin-right: var(--_3);
+    }
+    .mr-4 {
+        margin-right: var(--_4);
+    }
+    .mr-5 {
+        margin-right: var(--_5);
+    }
+    .mr-6 {
+        margin-right: var(--_6);
+    }
+    .mr-7 {
+        margin-right: var(--_7);
+    }
+    .mr-8 {
+        margin-right: var(--_8);
+    }
+    .mr-9 {
+        margin-right: var(--_9);
+    }
+    .mr-10 {
+        margin-right: var(--_10);
+    }
+
+
+    /** Margin Top */
+
+    .mt-0 {
+        margin-top: var(--_0);
+    }
+    .mt-1 {
+        margin-top: var(--_1);
+    }
+    .mt-2 {
+        margin-top: var(--_2);
+    }
+    .mt-3 {
+        margin-top: var(--_3);
+    }
+    .mt-4 {
+        margin-top: var(--_4);
+    }
+    .mt-5 {
+        margin-top: var(--_5);
+    }
+    .mt-6 {
+        margin-top: var(--_6);
+    }
+    .mt-7 {
+        margin-top: var(--_7);
+    }
+    .mt-8 {
+        margin-top: var(--_8);
+    }
+    .mt-9 {
+        margin-top: var(--_9);
+    }
+    .mt-10 {
+        margin-top: var(--_10);
+    }
+
+    /** Margins Y */
+
+
+    .my-0 {
+        margin-top: var(--_0);
+        margin-bottom: var(--_0);
+    }
+    .my-1 {
+        margin-top: var(--_1);
+        margin-bottom: var(--_1);
+    }
+    .my-2 {
+        margin-top: var(--_2);
+        margin-bottom: var(--_2);
+    }
+    .my-3 {
+        margin-top: var(--_3);
+        margin-bottom: var(--_3);
+    }
+    .my-4 {
+        margin-top: var(--_4);
+        margin-bottom: var(--_4);
+    }
+    .my-5 {
+        margin-top: var(--_5);
+        margin-bottom: var(--_5);
+    }
+    .my-6 {
+        margin-top: var(--_6);
+        margin-bottom: var(--_6);
+    }
+    .my-7 {
+        margin-top: var(--_7);
+        margin-bottom: var(--_7);
+    }
+    .my-8 {
+        margin-top: var(--_8);
+        margin-bottom: var(--_8);
+    }
+    .my-9 {
+        margin-top: var(--_9);
+        margin-bottom: var(--_9);
+    }
+    .my-10 {
+        margin-top: var(--_10);
+        margin-bottom: var(--_10);
+    }
+
+/**! Paddings */
+
+    /** Padding */
+
+    .p-0 {
+        padding: var(--_0);
+    }
+    .p-1 {
+        padding: var(--_1);
+    }
+    .p-2 {
+        padding: var(--_2);
+    }
+    .p-3 {
+        padding: var(--_3);
+    }
+    .p-4 {
+        padding: var(--_4);
+    }
+    .p-5 {
+        padding: var(--_5);
+    }
+    .p-6 {
+        padding: var(--_6);
+    }
+    .p-7 {
+        padding: var(--_7);
+    }
+    .p-8 {
+        padding: var(--_8);
+    }
+    .p-9 {
+        padding: var(--_9);
+    }
+    .p-10 {
+        padding: var(--_10);
+    }
+
+    /** Padding Right */
+
+    .pr-0 {
+        padding-right: var(--_0);
+    }
+    .pr-1 {
+        padding-right: var(--_1);
+    }
+    .pr-2 {
+        padding-right: var(--_2);
+    }
+    .pr-3 {
+        padding-right: var(--_3);
+    }
+    .pr-4 {
+        padding-right: var(--_4);
+    }
+    .pr-5 {
+        padding-right: var(--_5);
+    }
+    .pr-6 {
+        padding-right: var(--_6);
+    }
+    .pr-7 {
+        padding-right: var(--_7);
+    }
+    .pr-8 {
+        padding-right: var(--_8);
+    }
+    .pr-9 {
+        padding-right: var(--_9);
+    }
+    .pr-10 {
+        padding-right: var(--_10);
+    }
+
+    /** Padding Left */
+
+    .pl-0 {
+        padding-left: var(--_0);
+    }
+    .pl-1 {
+        padding-left: var(--_1);
+    }
+    .pl-2 {
+        padding-left: var(--_2);
+    }
+    .pl-3 {
+        padding-left: var(--_3);
+    }
+    .pl-4 {
+        padding-left: var(--_4);
+    }
+    .pl-5 {
+        padding-left: var(--_5);
+    }
+    .pl-6 {
+        padding-left: var(--_6);
+    }
+    .pl-7 {
+        padding-left: var(--_7);
+    }
+    .pl-8 {
+        padding-left: var(--_8);
+    }
+    .pl-9 {
+        padding-left: var(--_9);
+    }
+    .pl-10 {
+        padding-left: var(--_10);
+    }
+
+
+    /** Padding Y */
+
+    .py-0 {
+        padding-bottom: var(--_0);
+        padding-top: var(--_0);
+    }
+    .py-1 {
+        padding-bottom: var(--_1);
+        padding-top: var(--_1);
+    }
+    .py-2 {
+        padding-bottom: var(--_2);
+        padding-top: var(--_2);
+    }
+    .py-3 {
+        padding-bottom: var(--_3);
+        padding-top: var(--_3);
+    }
+    .py-4 {
+        padding-bottom: var(--_4);
+        padding-top: var(--_4);
+    }
+    .py-5 {
+        padding-bottom: var(--_5);
+        padding-top: var(--_5);
+    }
+    .py-6 {
+        padding-bottom: var(--_6);
+        padding-top: var(--_6);
+    }
+    .py-7 {
+        padding-bottom: var(--_7);
+        padding-top: var(--_7);
+    }
+    .py-8 {
+        padding-bottom: var(--_8);
+        padding-top: var(--_8);
+    }
+    .py-9 {
+        padding-bottom: var(--_9);
+        padding-top: var(--_9);
+    }
+    .py-10 {
+        padding-bottom: var(--_10);
+        padding-top: var(--_10);
+    }
+
+
+
+
+
+
+
+
+
+    .mw-100 {
+        max-width: 100%;
+    }
